# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档 for Aurora 项目",
        "title": "Aurora API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
...
        "/chat/conversations_list": {
            "get": {
                "description": "获取当前用户的所有对话列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取对话列表",
                "responses": {
                    "200": {
                        "description": "获取对话列表成功",
                        "schema": {
                            "$ref": "#/definitions/routes.conversationsListResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "获取对话列表失败",
                        "schema": {
                            "$ref": "#/definitions/routes.conversationsListResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/delete_conversation": {
            "post": {
                "description": "删除指定的对话及其相关消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "删除对话",
                "parameters": [
                    {
                        "description": "删除对话请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.deleteConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除对话成功",
                        "schema": {
                            "$ref": "#/definitions/routes.deleteConversationResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "删除对话失败",
                        "schema": {
                            "$ref": "#/definitions/routes.deleteConversationResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/delete_message": {
            "post": {
                "description": "删除指定的消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "删除消息",
                "parameters": [
                    {
                        "description": "删除消息请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.deleteMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除消息成功",
                        "schema": {
                            "$ref": "#/definitions/routes.deleteMessageResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "删除消息失败",
                        "schema": {
                            "$ref": "#/definitions/routes.deleteMessageResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/generate": {
            "post": {
                "description": "根据用户输入生成AI回复，使用SSE流式传输",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "生成AI回复",
                "parameters": [
                    {
                        "description": "生成请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.generateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "流式响应，返回AI生成内容",
                        "schema": {
                            "$ref": "#/definitions/routes.generateResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "生成失败",
                        "schema": {
                            "$ref": "#/definitions/routes.generateResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/messages_list": {
            "post": {
                "description": "获取指定对话的聊天记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取历史消息",
                "parameters": [
                    {
                        "description": "获取历史消息请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.messagesListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取历史消息成功",
                        "schema": {
                            "$ref": "#/definitions/routes.messagesListResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "获取历史消息失败",
                        "schema": {
                            "$ref": "#/definitions/routes.messagesListResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/new_conversation": {
            "get": {
                "description": "为当前用户创建一个新的对话",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "创建新对话",
                "responses": {
                    "200": {
                        "description": "创建新对话成功",
                        "schema": {
                            "$ref": "#/definitions/routes.newConversationResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "创建新对话失败",
                        "schema": {
                            "$ref": "#/definitions/routes.newConversationResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/share_messages": {
            "post": {
                "description": "将指定的消息ID列表创建为分享链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "分享消息",
                "parameters": [
                    {
                        "description": "分享消息请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.shareMessagesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建分享成功",
                        "schema": {
                            "$ref": "#/definitions/routes.shareMessagesResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "创建分享失败",
                        "schema": {
                            "$ref": "#/definitions/routes.shareMessagesResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/stop": {
            "post": {
                "description": "停止指定对话的AI生成过程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "停止生成",
                "parameters": [
                    {
                        "description": "停止生成请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.stopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "停止生成成功",
                        "schema": {
                            "$ref": "#/definitions/routes.stopResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "停止生成失败",
                        "schema": {
                            "$ref": "#/definitions/routes.stopResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/stt": {
            "post": {
                "description": "将语音转换为文字",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "语音转文字",
                "parameters": [
                    {
                        "description": "语音转文字请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.sttRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "语音转文字成功",
                        "schema": {
                            "$ref": "#/definitions/routes.sttResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "语音转文字失败",
                        "schema": {
                            "$ref": "#/definitions/routes.sttResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/thread_list": {
            "post": {
                "description": "获取当前用户的线程列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取线程列表",
                "responses": {
                    "200": {
                        "description": "获取线程列表成功",
                        "schema": {
                            "$ref": "#/definitions/routes.threadListResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "获取线程列表失败",
                        "schema": {
                            "$ref": "#/definitions/routes.threadListResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/tts": {
            "post": {
                "description": "将文字转换为语音",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "文字转语音",
                "parameters": [
                    {
                        "description": "文字转语音请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.ttsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文字转语音成功",
                        "schema": {
                            "$ref": "#/definitions/routes.ttsResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "文字转语音失败",
                        "schema": {
                            "$ref": "#/definitions/routes.ttsResponseFailed"
                        }
                    }
                }
            }
        },
        "/chat/{shareID}": {
            "get": {
                "description": "根据分享ID获取分享的消息内容",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取分享内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享ID",
                        "name": "shareID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取分享内容成功",
                        "schema": {
                            "$ref": "#/definitions/routes.loadShareMessagesResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "获取分享内容失败",
                        "schema": {
                            "$ref": "#/definitions/routes.loadShareMessagesResponseFailed"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "routes.conversationsListResponseFailed": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.conversationsListResponseSuccess": {
            "type": "object",
            "properties": {
                "conversations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.Conversation"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "routes.deleteConversationRequest": {
            "type": "object",
            "properties": {
                "conversationID": {
                    "type": "integer",
                    "example": 1234567890
                }
            }
        },
        "routes.deleteConversationResponseFailed": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.deleteConversationResponseSuccess": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "routes.deleteMessageRequest": {
            "type": "object",
            "properties": {
                "messageID": {
                    "type": "integer",
                    "example": 1234567890
                }
            }
        },
...
        "routes.deleteMessageResponseSuccess": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}